feat: support module: override key for multi-instance modules
Allows multiple config sections to share one Python implementation by
adding an optional 'module:' key to a config section. The framework
loads the specified module instead of inferring it from the section name.

- scripts/pivac-provider.py: check packages[p].get("module", p) when
  loading each module; fall back to section name if 'module:' absent
- pivac/ArduinoSensor.py: new consolidated implementation (was duplicated
  verbatim in ArduinoPSI.py and ArduinoThermPSI.py)
- pivac/ArduinoPSI.py, pivac/ArduinoThermPSI.py: replaced with thin
  wrappers that import from ArduinoSensor for backward compatibility
- config/config.yml.sample: add pivac.ArduinoPSI and pivac.ArduinoThermPSI
  sections with module: pivac.ArduinoSensor
- README.md: document module: override key in config section and in
  Adding a New Provider Module guide
